+++ src/Order/Events/OrderItemHasBeenPutOnHold.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the OrderItemHasBeenPutOnHold class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-03-14
+ *
+ */
+
+namespace Vanilo\Order\Events;
+
+// When the fulfillment status goes into ON_HOLD
+class OrderItemHasBeenPutOnHold extends BaseOrderItemEvent
+{
+}